requirements-option-b.txt
polars
pyarrow
sentence-transformers
faiss-cpu
fastapi
uvicorn
sqlite-utils
